refactor: detectCardCompany 함수 cardBrandUtils로 이동
